requirements.txt
scipy==1.1.0
Keras==2.1.6
websockets==7.0
dtaidistance==1.1.3
matplotlib==3.0.0
ciso8601==2.1.1
requests==2.18.4
pandas==0.23.0
numpy==1.14.3
setuptools==39.1.0
python_dateutil==2.8.0
scikit_learn==0.20.2